+++ public/scripts/service-worker.js
@@ -2,6 +2,9 @@
  * Course of action if serviceWorker is not supported
  * 
  * @returns {void}
+ * 
+ * @author rajat_m (https://medium.com/@rajat_m)
+ * @source https://medium.com/@rajat_m/what-are-service-workers-and-how-to-use-them-e993c1f497e6
  */
 function swUnsupported() {
     console.warn("Service Worker is not supported by browser.")
@@ -13,7 +16,7 @@ function swUnsupported() {
  * @returns {void}
  */
 function startServiceWorker() {
-    /* Guard-clause if service worker is not supported by the browser */
+    // Guard-clause if service worker is not supported by the browser
     if (!'serviceWorker' in navigator) return swUnsupported();
 
     window.addEventListener("load", () => {